Project_2_Purwadhika_DE/task_2/requirements.txt
python-dotenv==1.1.1
psycopg2-binary==2.9.10
psycopg==3.2.10
SQLAlchemy==2.0.43
beautifulsoup4==4.13.5
pandas==2.3.2
requests==2.32.5
pyarrow==21.0.0
scikit-learn==1.7.2
tqdm==4.67.1
openpyxl==3.1.5
